pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "strava-supercompensation"
version = "0.1.0"
description = "Automated Strava training analysis tool using supercompensation model"
authors = [
    { name = "Alex", email = "alex@example.com" }
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.31.0",
    "pandas>=2.1.0",
    "sqlalchemy>=2.0.0",
    "python-dotenv>=1.0.0",
    "click>=8.1.0",
    "aiohttp>=3.9.0",
    "authlib>=1.3.0",
    "numpy>=1.24.0",
    "rich>=13.7.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "flake8>=6.1.0",
    "mypy>=1.5.0",
]
viz = [
    "matplotlib>=3.7.0",
    "plotly>=5.17.0",
]
api = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
]

[project.scripts]
strava-super = "strava_supercompensation.cli:main"

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true